added routing to generatePDF
